+++ classes/attempt_report.php
@@ -380,6 +380,17 @@ class attempt_report {
                 .stackinputerror {
                     display: none !important;
                 }
+
+                /* Force code boxes to reflow to page width */
+                pre[class*='language-'] {
+                    overflow: visible !important;
+                    white-space: pre-wrap !important;
+                }
+
+                /* Remove padding from codebox comments to prevent them from drawing over student code */
+                code .token.comment {
+                    padding: 0.5rem !important;
+                }
             ");
             $dom->getElementsByTagName('head')[0]->appendChild($csshacksnode);
         }