+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version X.Y.Z (YYYYMMDDXX)
+
+- Force wrapping of long lines in code boxes to prevent overflowing out of page boundaries
+- Reduce padding of comment boxes within code boxes to prevent them from overlapping student code
+
+
 ## Version 1.0.0 (2025102700)
 
 - First stable release 🎉